+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## New version 1.1.0 - 31.08.2020 Shadowlands
+## New version 1.1.0 - 01.09.2020 Shadowlands
 
 ### Added in 1.1.0
 
 - Added option to select and manage whitelist for disenchanting
+- Added new macros for Shadowlands mass techniques
+- Added new herbs and ores from Shadowlands
 
 ### Notes